+++ CHANGELOG.md
@@ -6,8 +6,9 @@ Version 5.1-SNAPSHOT
 
 ADDED:
 
-- new ContractVerifier class 
-- new check methods in class EqualsHashCodeAndToStringCheck
+- new class ContractVerifier created
+- new annotation class Experimental created
+- new check methods in class EqualsHashCodeAndToStringCheck created
 
 Version 5
 -------------